@@ -2,8 +2,8 @@ package captionforpictures.kayushi07.com.captionforpictures.widgets;
 
 import android.content.Context;
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+import androidx.recyclerview.widget.RecyclerView;
 
 public class GridMarginDecoration extends RecyclerView.ItemDecoration {
 
